Move generic SHC functions and IPS resource extraction function to util file
